refactor(testing): consolidate testing modules
- Move groovy-testing → ext/testing (gvy.testing package)
- Rename groovy-test-utils → test-utils (gvy.testutils package)
- Update dependencies and imports across affected modules

Completes the groovy-* → gvy.* namespace migration.
