+++ ext/README.md
@@ -4,5 +4,6 @@ Extension modules providing LSP support for specific tools, libraries, and frame
 
 ## Modules
 
+- `testing/` - Test framework detection API (package: `gvy.testing`)
 - `spock/` - Spock testing framework support
 - `junit/` - JUnit 4 & 5 testing framework support
