+++ settings.gradle.kts
@@ -35,8 +35,8 @@ include("dsl:gdsl")
 include("dsl:dsld")
 include("build-tool")
 include("ext:spock")
-include("groovy-testing")
-include("groovy-test-utils")
+include("ext:testing")
+include("test-utils")
 include("ext:junit")
 include("jupyter:kernel-core")
 include("jupyter:kernels:groovy")
